pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["matfonseca <fonseca.196@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.1"
python-dotenv = "^1.0.0"
langserve = "^0.0.39"
fastapi = "^0.109.0"
openai = "^1.8.0"
sse-starlette = "^1.8.2"
langchain-community = "^0.0.13"
pydantic = "1.10.13"
langchain-openai = "^0.0.2"
beautifulsoup4 = "^4.12.2"
faiss-cpu = "^1.7.4"
uvicorn = "^0.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
